// src/routines/interfaces/routine.interface.ts
// Generated by https://quicktype.io

export interface FindAllRoutineResponse {
  id: string;
  intensity: number;
  estimatedMins: number;
  name: string;
  description: string;
  level: string;
  slug: string;
  types: string[];
  images: string[];
}

// Generated by https://quicktype.io

export interface FindOneRoutineResponse {
  id: string;
  intensity: number;
  estimatedMins: number;
  name: string;
  description: string;
  level: string;
  slug: string;
  types: string[];
  images: string[];
}

export interface Image {
  id: number;
  url: string;
}